% Generated by roxygen2 (4.0.2): do not edit by hand
\name{baseURL}
\alias{baseURL}
\title{Return base URL to SCB API (Function is depricated)}
\usage{
baseURL(version = "v1", lang = "sv", ...)
}
\arguments{
\item{version}{The version of SCB API to use. (Default: \code{v1})}

\item{lang}{The language (two letters) to use in the fetched data. (Default: \code{sv})}

\item{...}{Additional parameters. These are currently ignored.}
}
\description{
Function is depricated, please use \code{\link{base_url}()} instead.
}

