\name{checkForLevels}
\alias{checkForLevels}
\title{Function to silently test for existence of metadata and return TRUE or FALSE based on the result of that test.}
\usage{
  checkForLevels(url = baseURL())
}
\arguments{
  \item{url}{API url}
}
\description{
  Function to silently test for existence of metadata and
  return TRUE or FALSE based on the result of that test.
}

