\name{deparseLevels}
\alias{deparseLevels}
\title{Function to deparse an URL into the nodes into}
\usage{
  deparseLevels(place, returnDistance = FALSE)
}
\arguments{
  \item{place}{Location in hierarchy, in the form of a
  URL.}

  \item{returnDistance}{Whether to return only the distance
  (in nodes) from top node to URL. (Default: \code{FALSE})}
}
\description{
  Function to deparse an URL into the nodes into
}

