% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findData}
\alias{findData}
\title{Find and download data from SCB (Function is depricated)}
\usage{
findData(history = FALSE, ...)
}
\arguments{
\item{history}{keep the history when the function is running.}

\item{...}{further parameters. These are currently ignored.}
}
\description{
Function is depricated, please use \link{find_scb_data} instead.
}

