\name{findSCBdata}
\alias{findSCBdata}
\title{Wrapper function to simply find and download data from SCB to the current R session.}
\usage{
  findSCBdata(history = FALSE)
}
\arguments{
  \item{history}{keep the history when the function is
  running.}
}
\description{
  Wrapper function to simply find and download data from
  SCB to the current R session.
}
\seealso{
  \code{\link{scbGetMetadata}}, \code{\link{getData}}
}

