\docType{data}
\name{rSCBapi}
\alias{rSCBapi}
\title{rSCBapi}
\format{List of 7
 $ base_url   : chr "http://api.scb.se"
 $ api_name   : chr "OV0104"
 $ api_version: chr "v1"
 $ database_id: chr "doris"
 $ language   : chr "sv"
 $ levels     : chr "ssd"
 $ table_id   : chr "BE/BE0401/BE0401B/BefProgFoddaMedel10"}
\usage{
  scbQueryData
}
\description{
  A package for interfacing with the API of Statistics
  Sweden (Statistiska Centralbyrån)
}
\section{...}{

}
\keyword{datasets}

