% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scb_api_parameters}
\alias{scb_api_parameters}
\title{Get API - parameters for Statistics Swedens API}
\usage{
scb_api_parameters()
}
\description{
Wrapper function to get the api_parameters for statistics Sweden using \code{\link{api_parameters}}.
}
\examples{
scb_api_parameters()
}
\seealso{
\code{\link{api_parameters}}
}

