\name{findSCBdata}
\alias{findSCBdata}
\title{SCB API data browser}
\usage{
  findSCBdata(history = FALSE, ...)
}
\arguments{
  \item{history}{keep the history when the function is
  running.}

  \item{...}{further parameters. These are currently
  ignored.}
}
\description{
  A wrapper function to simply find and download data from
  the SCB API into the current R session.
}
\seealso{
  \code{\link{scbGetMetadata}}, \code{\link{scbGetData}}
}

