% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scbGetData}
\alias{scbGetData}
\title{Get data from a bottom node in SCB API (Function is depricated)}
\usage{
scbGetData(url, dims, clean = FALSE)
}
\arguments{
\item{url}{URL to get data from (it is usually sufficient to submit the base URL, supplied via the \code{\link{base_url}} function, and the name of the variable).}

\item{dims}{A list of dimensional parameters to filter data by. Note that values \emph{must} be submitted for all dimensions of the data. If you don't want to filter data, submit an asterisk in quotation marks ("*") instead of values for that dimension.}

\item{clean}{Clean and melt the data to R format.}
}
\description{
Function is depricated. Please use \code{\link{get_pxweb_data}} instead. To get api parameters, use \code{\link{scb_api_parameters}}.
}
\seealso{
\code{\link{get_pxweb_metadata}}, \code{\link{get_pxweb_dims}}, \code{\link{get_pxweb_levels}}
}

