% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scbGetDims}
\alias{scbGetDims}
\title{Get data dimensions form a bottom node in SCB API (Function is depricated)}
\usage{
scbGetDims(node, verbose = TRUE)
}
\arguments{
\item{node}{Bottom node to deparse into object}

\item{verbose}{Verbose output mode.}
}
\description{
Function is depricated. Please use \code{\link{get_pxweb_dims}} instead.
}
\seealso{
\code{\link{get_pxweb_metadata}}
}

