\name{scbGetDims}
\alias{scbGetDims}
\title{Get data dimensions form a bottom node in SCB API}
\usage{
  scbGetDims(url, verbose = TRUE)
}
\arguments{
  \item{url}{URL to get data from}

  \item{verbose}{Verbose output mode.}
}
\description{
  Deparse and reorder data form the metadata object for
  pretty output
}

