\name{scbGetDims}
\alias{scbGetDims}
\title{Get data dimensions form a bottom node in SCB API}
\usage{
scbGetDims(node, verbose = TRUE)
}
\arguments{
  \item{node}{Bottom node to deparse into object}

  \item{verbose}{Verbose output mode.}
}
\description{
Deparse and reorder data form the metadata object for
pretty output
}

