% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scbGetLevels}
\alias{scbGetLevels}
\title{Get levels from Statistics Sweden API node (Function is depricated)}
\usage{
scbGetLevels(baseURL, descriptions = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{baseURL}{URL to get levels in the API for.}

\item{descriptions}{Whether to include node descriptions with the list of node IDs. (default: \code{FALSE})}

\item{quiet}{Quiet mode. Whether to stop with an error if the input node does not contain any subnodes. If set to \code{TRUE}, the function will quietly return FALSE without any errors. (default: \code{FALSE})}

\item{...}{further argument to send to \code{get_pxweb_metadata()}}
}
\description{
Function is depricated, please use  \code{\link{get_pxweb_levels}} instead.
}

