% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scbGetMetadata}
\alias{scbGetMetadata}
\title{Get metadata (Function is depricated)}
\usage{
scbGetMetadata(path = NULL, node = NULL, topnodes = NULL, quiet = TRUE,
  ...)
}
\arguments{
\item{path}{URL to fetch metadata from. If left empty, the function constructs a URL from the \code{node} and \code{topnodes} arguments}

\item{node}{A string with the name of the node to fetch. This is ignored if \code{path} is supplied.}

\item{topnodes}{A string or list of strings with the names of the parent nodes of \code{node}. This is ignored if \code{path} is supplied.}

\item{quiet}{Quiet mode (never return a message to the user)}

\item{...}{Further arguments passed to  \code{\link{base_url}()}.}
}
\description{
Function is depricated, please use \code{\link{get_pxweb_metadata}()} instead.
}

